+++ heim-memory/src/sys/macos/mod.rs
@@ -3,9 +3,9 @@ mod swap;
 mod bindings;
 
 lazy_static::lazy_static! {
-    static ref PAGE_SIZE: usize = {
+    static ref PAGE_SIZE: u64 = {
         unsafe {
-            libc::sysconf(libc::_SC_PAGESIZE) as usize
+            libc::sysconf(libc::_SC_PAGESIZE) as u64
         }
     };
 }
